requirements.txt
fastapi==0.109.0
httpx==0.26.0
psycopg2-binary==2.9.9
pytest==8.0.0
pytest-cov==4.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
uvicorn==0.27.0